+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix parsing for mwld 2.7+ generated mapfiles.
+
 ## [2.9.3] - 2025-06-01
 
 ### Fixed